requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
pydantic-settings
httpx
apscheduler
passlib[bcrypt]
python-jose[cryptography]
celery
redis
